backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-speech
python-multipart
pydantic
openai
python-dotenv